+++ src/main/java/com/nexters/goalpanzi/application/firebase/event/handler/PushNotificationEventHandler.java
@@ -3,7 +3,7 @@ package com.nexters.goalpanzi.application.firebase.event.handler;
 import com.nexters.goalpanzi.application.firebase.TopicGenerator;
 import com.nexters.goalpanzi.application.mission.event.CompleteMissionEvent;
 import com.nexters.goalpanzi.application.mission.event.JoinMissionEvent;
-import com.nexters.goalpanzi.infrastructure.firebase.PushNotificationSender;
+import com.nexters.goalpanzi.infrastructure.firebase.PushMessageSender;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -13,21 +13,21 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-import static com.nexters.goalpanzi.domain.firebase.PushNotificationMessage.MISSION_COMPLETED;
-import static com.nexters.goalpanzi.domain.firebase.PushNotificationMessage.MISSION_JOINED;
+import static com.nexters.goalpanzi.domain.firebase.PushMessage.MISSION_COMPLETED;
+import static com.nexters.goalpanzi.domain.firebase.PushMessage.MISSION_JOINED;
 
 @Slf4j
 @RequiredArgsConstructor
 @Component
 public class PushNotificationEventHandler {
 
-    private final PushNotificationSender pushNotificationSender;
+    private final PushMessageSender pushMessageSender;
 
     @Async
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     void handleJoinMissionEvent(final JoinMissionEvent event) {
-        pushNotificationSender.sendIndividualData(
+        pushMessageSender.sendIndividualData(
                 MISSION_JOINED.getTitle(),
                 MISSION_JOINED.getBody(event.nickname()),
                 event.deviceToken(),
@@ -41,7 +41,7 @@ public class PushNotificationEventHandler {
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     void handleCompleteMissionEvent(final CompleteMissionEvent event) {
         String topic = TopicGenerator.getTopic(event.missionId());
-        pushNotificationSender.sendGroupData(
+        pushMessageSender.sendGroupData(
                 MISSION_COMPLETED.getTitle(),
                 MISSION_COMPLETED.getBody(),
                 topic,
